+++ internal/models/models.go
@@ -5,10 +5,12 @@ type Conf struct {
 	Host     string
 	Port     string
 	Theme    string
+	Color    string
 	Icon     string
 	ConfPath string
 	YamlPath string
 	LogPath  string
+	NodePath string
 	Quit     chan bool
 }