+++ internal/web/config.go
@@ -12,7 +12,7 @@ func configHandler(w http.ResponseWriter, r *http.Request) {
 
 	guiData.Config = AppConfig
 
-	guiData.Themes = []string{"cerulean", "cosmo", "cyborg", "darkly", "flatly", "journal", "litera", "lumen", "lux", "materia", "minty", "morph", "pulse", "quartz", "sandstone", "simplex", "sketchy", "slate", "solar", "spacelab", "superhero", "united", "vapor", "yeti", "zephyr"}
+	guiData.Themes = []string{"cerulean", "cosmo", "cyborg", "darkly", "emerald", "flatly", "grass", "grayscale", "journal", "litera", "lumen", "lux", "materia", "minty", "morph", "ocean", "pulse", "quartz", "sand", "sandstone", "simplex", "sketchy", "slate", "solar", "spacelab", "superhero", "united", "vapor", "wood", "yeti", "zephyr"}
 
 	execTemplate(w, "config", guiData)
 }
@@ -22,6 +22,9 @@ func saveConfigHandler(w http.ResponseWriter, r *http.Request) {
 	AppConfig.Host = r.FormValue("host")
 	AppConfig.Port = r.FormValue("port")
 	AppConfig.Theme = r.FormValue("theme")
+	AppConfig.Color = r.FormValue("color")
+	AppConfig.NodePath = r.FormValue("nodepath")
+
 	conf.Write(AppConfig.ConfPath, AppConfig)
 
 	http.Redirect(w, r, r.Header.Get("Referer"), 302)